start bundling mvdan.cc/sh/v3/pattern
	bundle -o pattern_bundle.go -prefix pattern mvdan.cc/sh/v3/pattern

This makes the module flat and with no dependencies, at the cost of just
a few hundred lines of bundled code.

Most importantly, we want to use editorconfig in the sh module for
shfmt, so this allows avoiding cyclic module dependencies.

Note that the sh/v3 module must exist in go.mod for the bundle command
to work.