refactor regexp caching to be less racy
Keeping a mutex inside Section was a bad idea, since we store and use
that struct without a pointer, hence copying the mutex too.

Instead, do the caching at the Query level. This also makes sense, since
we allow the user to configure whether or not they want any caching. If
no caching is wanted, this also makes the data structures leaner.

This change makes the API a bit nicer, and 'go vet' happy.